pyproject.toml
[project]
name = "analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fyers-apiv3>=3.1.12",
    "grpcio>=1.60.0",
    "grpcio-reflection>=1.60.0",
    "grpcio-tools>=1.60.0",
    "mcp>=1.0.0",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.2",
    "requests>=2.31.0",
    "truststore>=0.10.0",
    "tzdata>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["scheduler", "notify", "auth"]
python_files = "test_*.py"
